# util/tracing: reduce allocations in EnsureForkSpan