This patch pre-allocates capacity for the options slice in
EnsureForkSpan. Before, we were reallocating the slice storage
repeatedly, as it was growing from 0 to 1 and from 1 to 2 elements.

This patch also simplifies, and slighly speeds up, EnsureChildSpan. This
guy was using a sync.Pool for its options slice. That is not necessary,
since the slice storage does not escape.

In the benchmarks below, detached-child=false is EnsureChildSpan and
detached-child=true is EnsureForkSpan.

> make bench PKG=./pkg/util/tracing BENCHES=BenchmarkSpanCreation TESTFLAGS="-count=5 -cpu=1,2,4,8,16"
name                                  old time/op    new time/op    delta
SpanCreation/detached-child=false       1.70µs ± 1%    1.61µs ± 1%   -5.06%  (p=0.008 n=5+5)
SpanCreation/detached-child=false-2     1.17µs ± 7%    1.04µs ± 3%  -10.70%  (p=0.008 n=5+5)
SpanCreation/detached-child=false-4      958ns ± 2%     893ns ±10%   -6.76%  (p=0.008 n=5+5)
SpanCreation/detached-child=false-8     1.11µs ±38%    0.95µs ± 1%     ~     (p=0.310 n=5+5)
SpanCreation/detached-child=false-16     925ns ± 1%     919ns ± 1%     ~     (p=0.286 n=4+5)
SpanCreation/detached-child=true        2.27µs ± 2%    1.88µs ± 0%  -17.27%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-2      1.82µs ±10%    1.33µs ±22%  -26.81%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-4      1.29µs ± 2%    1.02µs ± 1%  -21.35%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-8      1.16µs ± 1%    0.91µs ± 5%  -21.56%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-16     1.14µs ± 2%    0.97µs ± 1%  -15.08%  (p=0.008 n=5+5)

name                                  old alloc/op   new alloc/op   delta
SpanCreation/detached-child=false         288B ± 0%      288B ± 0%     ~     (all equal)
SpanCreation/detached-child=false-2       288B ± 0%      288B ± 0%     ~     (all equal)
SpanCreation/detached-child=false-4       288B ± 0%      288B ± 0%     ~     (all equal)
SpanCreation/detached-child=false-8       288B ± 0%      288B ± 0%     ~     (all equal)
SpanCreation/detached-child=false-16      288B ± 0%      288B ± 0%     ~     (all equal)
SpanCreation/detached-child=true          768B ± 0%      288B ± 0%  -62.50%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-2        768B ± 0%      288B ± 0%  -62.50%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-4        768B ± 0%      288B ± 0%  -62.50%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-8        768B ± 0%      288B ± 0%  -62.50%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-16       769B ± 0%      288B ± 0%  -62.55%  (p=0.008 n=5+5)

name                                  old allocs/op  new allocs/op  delta
SpanCreation/detached-child=false         6.00 ± 0%      6.00 ± 0%     ~     (all equal)
SpanCreation/detached-child=false-2       6.00 ± 0%      6.00 ± 0%     ~     (all equal)
SpanCreation/detached-child=false-4       6.00 ± 0%      6.00 ± 0%     ~     (all equal)
SpanCreation/detached-child=false-8       6.00 ± 0%      6.00 ± 0%     ~     (all equal)
SpanCreation/detached-child=false-16      6.00 ± 0%      6.00 ± 0%     ~     (all equal)
SpanCreation/detached-child=true          16.0 ± 0%       6.0 ± 0%  -62.50%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-2        16.0 ± 0%       6.0 ± 0%  -62.50%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-4        16.0 ± 0%       6.0 ± 0%  -62.50%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-8        16.0 ± 0%       6.0 ± 0%  -62.50%  (p=0.008 n=5+5)
SpanCreation/detached-child=true-16       16.0 ± 0%       6.0 ± 0%  -62.50%  (p=0.008 n=5+5)

Release note: None